Fix compilation

+++ ch32v305/src/main.c
@@ -174,7 +174,7 @@ bool tud_vendor_control_xfer_cb(uint8_t rhport, uint8_t stage, tusb_control_requ
   return false;
 }
 
-void tud_vendor_rx_cb(uint8_t itf, uint8_t const* buffer, uint16_t bufsize) {
+void tud_vendor_rx_cb(uint8_t itf, uint8_t const* buffer, uint32_t bufsize) {
   (void) itf;
 
   // if using RX buffered is enabled, we need to flush the buffer to make room for new data
